+++ src/main/java/com/sparta/instahub/admin/service/AdminService.java
@@ -3,8 +3,8 @@ package com.sparta.instahub.admin.service;
 import com.sparta.instahub.auth.entity.User;
 import com.sparta.instahub.post.entity.Post;
 import com.sparta.instahub.post.repository.PostRepository;
-import com.sparta.instahub.user.entity.UserRole;
-import com.sparta.instahub.user.entity.UserStatus;
+import com.sparta.instahub.auth.entity.entity.UserRole;
+import com.sparta.instahub.auth.entity.entity.UserStatus;
 import com.sparta.instahub.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
