+++ src/main/java/com/sparta/instahub/auth/entity/entity/UserRole.java
@@ -1,0 +1,6 @@
+package com.sparta.instahub.auth.entity.entity;
+
+public enum UserRole {
+    ADMIN,
+    USER
+}
