--- src/main/java/com/sparta/instahub/user/entity/UserRole.java
@@ -1,6 +1,0 @@
-package com.sparta.instahub.user.entity;
-
-public enum UserRole {
-    ADMIN,
-    USER
-}
